requirements.txt
python-dotenv
spotipy
path
config
apscheduler